+++ src/modules/index.ts
@@ -1,3 +1,4 @@
 export * from './EventHandler';
 export * from './LocalStorage';
 export * from './Observer';
+export * from './Message';
